Added StockController CRUD actions

+++ Controllers/StockTransactionsController.cs
@@ -22,7 +22,9 @@ namespace StockPortfolio.Controllers
         // GET: StockTransactions
         public async Task<IActionResult> Index()
         {
-            var portfolioContext = _context.StockTransactions.Include(s => s.Stock);
+            var portfolioContext = _context.StockTransactions
+                .Include(s => s.Stock)
+                .OrderByDescending(s => s.TransactionDateTime);
             return View(await portfolioContext.ToListAsync());
         }
 
@@ -36,19 +38,21 @@ namespace StockPortfolio.Controllers
 
             var stockTransaction = await _context.StockTransactions
                 .Include(s => s.Stock)
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
+
             if (stockTransaction == null)
             {
                 return NotFound();
             }
-
             return View(stockTransaction);
         }
+
 
         // GET: StockTransactions/Create
         public IActionResult Create()
         {
-            ViewData["StockID"] = new SelectList(_context.Stocks, "ID", "ID");
+            ViewData["StockID"] = new SelectList(_context.Stocks, "ID", "CompanyName");
             return View();
         }
 
@@ -57,17 +61,25 @@ namespace StockPortfolio.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,StockID,TransactionDateTime,TransactionType,Price,Quantity")] StockTransaction stockTransaction)
+        public async Task<IActionResult> Create([Bind("StockID,TransactionDateTime,TransactionType,Price,Quantity")] StockTransaction stockTransaction)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(stockTransaction);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(stockTransaction);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
             }
-            ViewData["StockID"] = new SelectList(_context.Stocks, "ID", "ID", stockTransaction.StockID);
+            catch (DbUpdateException ex)
+            {
+                ModelState.AddModelError("", "Unable to save changes.");
+            }
+            ViewData["StockID"] = new SelectList(_context.Stocks, "ID", "CompanyName", stockTransaction.StockID);
             return View(stockTransaction);
         }
+
 
         // GET: StockTransactions/Edit/5
         public async Task<IActionResult> Edit(int? id)
@@ -82,64 +94,64 @@ namespace StockPortfolio.Controllers
             {
                 return NotFound();
             }
-            ViewData["StockID"] = new SelectList(_context.Stocks, "ID", "ID", stockTransaction.StockID);
+            ViewData["StockID"] = new SelectList(_context.Stocks, "ID", "CompanyName", stockTransaction.StockID);
             return View(stockTransaction);
         }
+
 
         // POST: StockTransactions/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,StockID,TransactionDateTime,TransactionType,Price,Quantity")] StockTransaction stockTransaction)
-        {
-            if (id != stockTransaction.ID)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(stockTransaction);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!StockTransactionExists(stockTransaction.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["StockID"] = new SelectList(_context.Stocks, "ID", "ID", stockTransaction.StockID);
-            return View(stockTransaction);
-        }
-
-        // GET: StockTransactions/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> EditPost(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
+            var stockTransactionToUpdate = await _context.StockTransactions
+                .FirstOrDefaultAsync(s => s.ID == id);
+            if (await TryUpdateModelAsync<StockTransaction>(
+                stockTransactionToUpdate, "", s => s.StockID, s => s.TransactionDateTime, s => s.TransactionType, s => s.Price, s => s.Quantity))
+            {
+                try
+                {
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                catch (DbUpdateException ex)
+                {
+                    ModelState.AddModelError("", "Unable to save changes.");
+                }
+            }
+            return View(stockTransactionToUpdate);
+        }
 
+
+        // GET: StockTransactions/Delete/5
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
             var stockTransaction = await _context.StockTransactions
+                .AsNoTracking()
                 .Include(s => s.Stock)
-                .FirstOrDefaultAsync(m => m.ID == id);
+                .FirstOrDefaultAsync(s => s.ID == id);
             if (stockTransaction == null)
             {
                 return NotFound();
             }
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] = "Delete failed.";
+            }
 
             return View(stockTransaction);
         }
+
 
         // POST: StockTransactions/Delete/5
         [HttpPost, ActionName("Delete")]
@@ -147,10 +159,23 @@ namespace StockPortfolio.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var stockTransaction = await _context.StockTransactions.FindAsync(id);
-            _context.StockTransactions.Remove(stockTransaction);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (stockTransaction == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                _context.StockTransactions.Remove(stockTransaction);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException ex)
+            {
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
+
+
 
         private bool StockTransactionExists(int id)
         {
